requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
billiard==4.2.1
cattrs==24.1.2
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
flatbuffers==25.2.10
idna==3.10
kombu==5.4.2
numpy==2.2.3
openmeteo_requests==1.3.0
openmeteo_sdk==1.19.0
pandas==2.2.3
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg==3.2.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
url-normalize==1.4.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
